+++ packages/web3-eth-ens/src/config.js
@@ -22,7 +22,7 @@
 
 /**
  * Source: https://docs.ens.domains/ens-deployments
- * EnergyWebChain source: https://energyweb.atlassian.net/wiki/spaces/EWF/pages/555745281/Using+ENS
+ * EnergyWebChain/Volta source: https://energyweb.atlassian.net/wiki/spaces/EWF/pages/555745281/Using+ENS
  *
  * @type {{addresses: {main: string, rinkeby: string, goerli: string, ropsten: string}}}
  */
@@ -33,6 +33,7 @@ var config = {
         rinkeby: "0x00000000000C2E074eC69A0dFb2997BA6C7d2e1e",
         goerli: "0x00000000000C2E074eC69A0dFb2997BA6C7d2e1e",
         energyWebChain: "0x0A6d64413c07E10E890220BBE1c49170080C6Ca0",
+        volta: "0xd7CeF70Ba7efc2035256d828d5287e2D285CD1ac",
     },
     // These ids obtained at ensdomains docs:
     // https://docs.ens.domains/contract-developer-guide/writing-a-resolver